🎨 fix: 버킷리스트 통신 완료, 컴포넌트 분리
