+++ client/data/todos.json
@@ -1,14 +1,19 @@
 {
   "todos": [
     {
-      "title": "잘 된다잉",
+      "title": "잘 된다.",
       "completed": false,
       "id": 1
     },
     {
-      "title": "patch가 잘 안된다잉..",
+      "title": "patch 너무너무 잘된다.",
       "completed": false,
       "id": 2
+    },
+    {
+      "title": "잘 된다잉~",
+      "completed": false,
+      "id": 3
     }
   ]
 }
