+++ client/src/components/Modal/CheckList/CheckList.js
@@ -1,8 +1,5 @@
 import styled from "styled-components";
 import axios from "axios";
-import { useEffect, useState } from "react";
-import { AiOutlinePlusCircle } from "react-icons/ai";
-// import { useParams } from "react-router-dom";
 export const InputContainer = styled.div`
   width: ${(props) => (props.width ? props.width : "350px")};
   height: ${(props) => (props.height ? props.height : "58px")};
@@ -75,18 +72,6 @@ export const ButtonContainer = styled.div`
   align-items: center;
 `;
 
-export const Update = styled.button`
-  cursor: pointer;
-  border: none;
-  background-color: #ffffff;
-  color: #535353;
-  width: 39px;
-`;
-export const Box = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
 export const MintLineButton2 = styled.button`
   height: 30px;
   width: ${(props) => (props.width ? props.width : "auto")};
@@ -121,50 +106,7 @@ export const MintButton2 = styled.button`
   }
 `;
 
-export const CheckList = ({ input, setInput, todos, setTodos, editTodo, setEditTodo, completed }) => {
-  const [isEdit, setIsEdit] = useState(false);
-
-  useEffect(() => {
-    if (editTodo) {
-      setInput(editTodo.title);
-    } else {
-      setInput("");
-    }
-  }, [setInput, editTodo]);
-
-  const onInputChange = (event) => {
-    setInput(event.target.value);
-  };
-
-  const onFormSubmit = async (event) => {
-    event.preventDefault();
-    setTodos([...todos, { title: input, completed: completed }]);
-    setInput("");
-    // else {
-    //   updateTodo(input, editTodo.id, editTodo.completed);
-    // }
-    // useEffect(() => {
-    const todoPost = {
-      title: input,
-      completed: false,
-    };
-    await axios
-      .post("http://localhost:4000/todos", todoPost)
-      .then((res) => console.log(res))
-      .then((err) => console.log(err));
-
-    await axios.get("http://localhost:4000/todos").then((result) => {
-      setTodos(result.data);
-    });
-    // }, []);
-  };
-
-  const updateTodo = (title, id) => {
-    const newTodo = todos.map((todo) => (todo.id === id ? { title, id, completed } : todo));
-    setTodos(newTodo);
-    setEditTodo("");
-  };
-
+export const CheckList = ({ todos, setTodos, setEditTodo, isEdit, setIsEdit }) => {
   const handleComplete = (todo) => {
     setTodos(
       todos.map((item) => {
@@ -186,23 +128,6 @@ export const CheckList = ({ input, setInput, todos, setTodos, editTodo, setEditT
       .then((err) => console.log("res1", err));
   };
 
-  const updateHandler = () => {
-    updateTodo(input, editTodo.id, editTodo.completed);
-    // console.log("patchid", editTodo.id);
-
-    const patch2 = {
-      title: input,
-      completed: completed,
-    };
-
-    axios
-      .patch(`http://localhost:4000/todos/` + editTodo.id, patch2)
-      .then((res) => console.log(res))
-      .then((err) => console.log("res2", err));
-
-    setIsEdit(!isEdit);
-  };
-
   const handleEdit = ({ id }) => {
     const findTodo = todos.find((todo) => todo.id === id);
     setEditTodo(findTodo);
@@ -216,51 +141,31 @@ export const CheckList = ({ input, setInput, todos, setTodos, editTodo, setEditT
   };
 
   return (
-    <>
-      <Box>
-        <InputContainer width="424px" height="36px">
-          <form onSubmit={onFormSubmit} className="form">
-            <Input type="text" value={input} required onChange={onInputChange} />
-            <Button type="submit">
-              <AiOutlinePlusCircle className="add" color="5E5E5E" size="18" />
-            </Button>
-          </form>
-        </InputContainer>
-        {isEdit ? (
-          <Update color="5E5E5E" size="18" onClick={(todos) => updateHandler(todos)}>
-            저장
-          </Update>
-        ) : null}
-      </Box>
-      <Test>
-        {todos.map((todo, idx) => (
-          <li className="list-item" key={idx}>
-            <Block2>
-              <button onClick={() => handleComplete(todo)} className="complete-icon">
-                {todo.completed ? "🔳" : "⬜"}
-              </button>
-              <input
-                type="text"
-                value={todo.title}
-                // 완료 시 밑줄 그어짐
-                className={`list ${todo.completed ? "complete" : ""}`}
-                onChange={(e) => e.preventDefault()}
-              />
-            </Block2>
-            <ButtonContainer>
-              <MintLineButton2 width="50px" height="20px" onClick={() => handleEdit(todo)}>
-                수정
-              </MintLineButton2>
-              <MintButton2 width="50px" height="20px" onClick={() => handleDelete(todo)}>
-                삭제
-              </MintButton2>
-            </ButtonContainer>
-          </li>
-        ))}
-      </Test>
-    </>
-    //   <div>
-    //   <BiCheckSquare className="check-icon" size={24} />
-    // </div>
+    <Test>
+      {todos.map((todo, idx) => (
+        <li className="list-item" key={idx}>
+          <Block2>
+            <button onClick={() => handleComplete(todo)} className="complete-icon">
+              {todo.completed ? "🔳" : "⬜"}
+            </button>
+            <input
+              type="text"
+              value={todo.title}
+              // 완료 시 밑줄 그어짐
+              className={`list ${todo.completed ? "complete" : ""}`}
+              onChange={(e) => e.preventDefault()}
+            />
+          </Block2>
+          <ButtonContainer>
+            <MintLineButton2 width="50px" height="20px" onClick={() => handleEdit(todo)}>
+              수정
+            </MintLineButton2>
+            <MintButton2 width="50px" height="20px" onClick={() => handleDelete(todo)}>
+              삭제
+            </MintButton2>
+          </ButtonContainer>
+        </li>
+      ))}
+    </Test>
   );
 };
